# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for BMT Manta Dashboard Service",
        "title": "BMT Manta Dashboard Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/api/heatmap": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve people count data by hour for the specified date",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "summary"
                ],
                "summary": "Get heatmap data by time period",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Date to retrieve data for (format YYYY-MM-DD). If not specified, current date will be used.",
                        "name": "date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.HeatmapData"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized (invalid or missing API key)",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/person-stats": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve statistics about new vs. returning people for the specified date",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "summary"
                ],
                "summary": "Get new vs. returning person statistics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Date to retrieve data for (format YYYY-MM-DD). If not specified, current date will be used.",
                        "name": "date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PersonStats"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized (invalid or missing API key)",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/summary": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve total, new, and returning people counts for the specified date",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "summary"
                ],
                "summary": "Get daily summary statistics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Date to retrieve data for (format YYYY-MM-DD). If not specified, current date will be used.",
                        "name": "date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DailySummary"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized (invalid or missing API key)",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "models.DailySummary": {
            "type": "object",
            "properties": {
                "date": {
                    "description": "Date in YYYY-MM-DD format",
                    "type": "string"
                },
                "new": {
                    "description": "Count of new people",
                    "type": "integer"
                },
                "repeat": {
                    "description": "Count of repeat visitors",
                    "type": "integer"
                },
                "total": {
                    "description": "Total count of people",
                    "type": "integer"
                }
            }
        },
...
        "models.HeatmapData": {
            "type": "object",
            "properties": {
                "count": {
                    "description": "Number of people detected during this hour",
                    "type": "integer"
                },
                "hour": {
                    "description": "Hour of the day in 24-hour format (e.g. \"08:00\", \"14:00\")",
                    "type": "string"
                }
            }
        },
...
        "models.PersonStats": {
            "type": "object",
            "properties": {
                "new": {
                    "description": "Count of new people",
                    "type": "integer"
                },
                "repeat": {
                    "description": "Count of repeat visitors",
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}